# backend/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "FamilyBudget",
    "contact": {
      "name": "Andriy Tryndey",
      "email": "atrundey@gmail.com"
    }
  },
...
  "paths": {
...
    "/transactions": {
      "post": {
        "tags": ["transactions"],
        "summary": "Add new transaction to store",
        "operationId": "createTransaction",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "default": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "sum",
            "in": "formData",
            "default": "100",
            "required": true,
            "type": "number"
          },
          {
            "name": "category",
            "in": "formData",
            "default": "travel",
            "required": true,
            "type": "string"
          },
          {
            "name": "kind",
            "in": "formData",
            "default": "food",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction was created",
            "schema": {
              "$ref": "#/definitions/singleTransaction"
            }
          },
          "400": {
            "description": "Just an error",
            "schema": {
              "$ref": "#/definitions/commonError"
            }
          }
        }
      },
      "get": {
        "tags": ["transactions"],
        "summary": "Pull out all transactions from the store",
        "operationId": "getAllTransaction",
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "default": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "default": 1,
            "required": false,
            "type": "number"
          },
          {
            "name": "perPage",
            "in": "query",
            "default": 20,
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "transactions was found",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/singleTransaction"
              }
            }
          },
          "400": {
            "description": "Just an error",
            "schema": {
              "$ref": "#/definitions/commonError"
            }
          }
        }
      }
    },
    "/transactions/{transaction_id}": {
      "get": {
        "tags": ["transactions"],
        "summary": "Get transaction from the store",
        "operationId": "getTransaction",
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "default": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "transaction was found",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/singleTransaction"
              }
            }
          },
          "400": {
            "description": "Just an error",
            "schema": {
              "$ref": "#/definitions/commonError"
            }
          }
        }
      },
      "put": {
        "tags": ["transactions"],
        "summary": "Change transaction from the store",
        "operationId": "putTransaction",
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "default": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "sum",
            "in": "formData",
            "default": 250,
            "required": false,
            "type": "number"
          },
          {
            "name": "category",
            "in": "formData",
            "default": "products",
            "required": false,
            "type": "string"
          },
          {
            "name": "kind",
            "in": "formData",
            "default": "clothes",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "transaction was changed",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/singleTransaction"
              }
            }
          },
          "400": {
            "description": "Just an error",
            "schema": {
              "$ref": "#/definitions/commonError"
            }
          }
        }
      },
      "delete": {
        "tags": ["transactions"],
        "summary": "Delete transaction from the store",
        "operationId": "deleteTransaction",
        "description": "",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "default": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "transaction was delete"
          },
          "400": {
            "description": "Just an error",
            "schema": {
              "$ref": "#/definitions/commonError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "singleTransaction": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "12345@ss",
          "description": "MongoDB ID"
        },
        "sum": {
          "type": "number",
          "example": 100
        },
        "category": {
          "type": "string",
          "example": "travel"
        },
        "kind": {
          "type": "string",
          "example": "food"
        },
        "full_name_user": {
          "type": "string",
          "example": "Nazar Nazarov"
        },
        "user_id": {
          "type": "string",
          "example": "23432@SS",
          "description": "MongoDB ID"
        }
      }
    },
    "commonError": {
      "type": "object",
      "properties": {
        "msg": {
          "type": "string",
          "example": "Something went wrong"
        },
        "code": {
          "type": "number",
          "example": 400
        }
      }
    }
  }
}